+++ nautilus_core/core/src/lib.rs
@@ -16,7 +16,6 @@
 pub mod correctness;
 pub mod cvec;
 pub mod datetime;
-pub mod enums;
 pub mod parsing;
 pub mod string;
 pub mod time;