+++ nautilus_core/model/src/orderbook/order.rs
@@ -20,20 +20,20 @@ use crate::types::quantity::Quantity;
 
 #[repr(C)]
 #[derive(Debug)]
-pub struct Order {
+pub struct BookOrder {
     pub price: Price,
     pub size: Quantity,
     pub side: OrderSide,
-    pub id: u64,
+    pub order_id: u64,
 }
 
-impl Order {
-    pub fn new(price: Price, size: Quantity, side: OrderSide, id: u64) -> Self {
-        Order {
+impl BookOrder {
+    pub fn new(price: Price, size: Quantity, side: OrderSide, order_id: u64) -> Self {
+        BookOrder {
             price,
             size,
             side,
-            id,
+            order_id,
         }
     }
 
@@ -42,7 +42,7 @@ impl Order {
     }
 }
 
-impl From<Vec<&str>> for Order {
+impl From<Vec<&str>> for BookOrder {
     fn from(vec: Vec<&str>) -> Self {
         assert_eq!(vec.len(), 4);
         let price = Price::from(vec[0]);
@@ -52,7 +52,7 @@ impl From<Vec<&str>> for Order {
             "S" => OrderSide::Sell,
             _ => panic!("Cannot parse side, was {}", vec[2]),
         };
-        Order::new(price, size, side, 0)
+        BookOrder::new(price, size, side, 0)
     }
 }
 
@@ -62,17 +62,17 @@ impl From<Vec<&str>> for Order {
 #[cfg(test)]
 mod tests {
     use crate::enums::OrderSide;
-    use crate::orderbook::order::Order;
+    use crate::orderbook::order::BookOrder;
     use crate::types::price::Price;
     use crate::types::quantity::Quantity;
 
     #[test]
     fn test_order_from_str_vec() {
         let input = vec!["1.00000", "100", "B", "123"];
-        let order = Order::from(input);
+        let order = BookOrder::from(input);
         assert_eq!(order.price, Price::new(1.0, 0));
         assert_eq!(order.size, Quantity::new(100.0, 0));
         assert_eq!(order.side, OrderSide::Buy);
-        assert_eq!(order.id, 0);
+        assert_eq!(order.order_id, 0);
     }
 }